Change way translations are stored
In order to be translated by Weblate, the translations can't be stored in a
string-array. We now store them in strings, with the synonyms separated by a
comma (,).

An example of this change can be found in app/src/main/res/values/verbs.xml.
The other verbs.xml files should be converted too.